Moved jumper image to separate file jumpImg.py
