+++ jumper.py
@@ -4,37 +4,7 @@ Program: Jumpman
 Authors: Daniel Jones(Lead), Steven Buechele, Eric Woll, Tebakaro Tione
 """
 
-
-# create image of jumping man
-# image gets shorter with each incorrect guess from main
-class jumpImg:
-    def __init__(self):
-        # Sets self.image as default picture of man jumping stored 
-        self.image = [
-        " ___ ",
-        "/___\\",
-        "\   /",
-        "  O  ",
-        " /|\ ",
-        " / \ "]
-        # Creates self.mistakes variable to keep track of number of bad guesses
-        self.mistakes = 0
-
-    '''
-    Checks number of mistakes
-    If mistakes > 2, change image head from 'O' to 'X'
-    Prints image line by line, skips 1 line per mistake
-    '''
-    def picture(self):
-        running = True
-        while running:
-            if self.mistakes > 2:
-                self.image[3] = "  X  "
-                running = False
-            for i in range(self.mistakes, 6):
-                print(self.image[i])
-            print()
-
+from jumpImg import *
 
 class PlayerInput:
     '''Retrieves user input and returns a single character'''
